Added better error handling and responce codes.

+++ src/SFA.DAS.AODP.Data/Repositories/PageRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using SFA.DAS.AODP.Data.Entities;
+using SFA.DAS.AODP.Data.Exceptions;
 using SFA.DAS.AODP.Infrastructure.Context;
 
 namespace SFA.DAS.AODP.Data.Repositories;
@@ -12,23 +13,60 @@ public class PageRepository : IPageRepository
         _context = context;
     }
 
+    /// <summary>
+    /// Gets all pages for a given section. 
+    /// Does not check if the section Id is valid, retuns an empty list if so.  
+    /// </summary>
+    /// <param name="sectionId"></param>
+    /// <returns></returns>
     public async Task<List<Page>> GetPagesForSectionAsync(Guid sectionId)
     {
         return await _context.Pages.Where(v => v.SectionId == sectionId).ToListAsync();
     }
 
-    public async Task<Page?> GetPageByIdAsync(Guid pageId)
+    /// <summary>
+    /// Gets a page with a given Id, throws if no page is found with the given Id. 
+    /// </summary>
+    /// <param name="pageId"></param>
+    /// <returns></returns>
+    /// <exception cref="RecordNotFoundException"></exception>
+    public async Task<Page> GetPageByIdAsync(Guid pageId)
     {
-        return await _context.Pages.FirstOrDefaultAsync(v => v.Id == pageId);
+        var res = await _context.Pages.FirstOrDefaultAsync(v => v.Id == pageId);
+        if (res is null)
+            throw new RecordNotFoundException(pageId);
+
+        return res;
     }
 
+    /// <summary>
+    /// Creates a page on a section, throws if no linked section is found, 
+    /// or if the linked form version isn't in draft status. 
+    /// </summary>
+    /// <param name="page"></param>
+    /// <returns></returns>
+    /// <exception cref="NoForeignKeyException"></exception>
+    /// <exception cref="RecordLockedException"></exception>
     public async Task<Page> Create(Page page)
     {
+        if (!await _context.Sections.AnyAsync(v => v.Id == page.SectionId))
+            throw new NoForeignKeyException(page.SectionId);
+
+        if (!await _context.Sections.AnyAsync(v => v.Id == page.SectionId && v.FormVersion.Status != FormStatus.Draft))
+            throw new RecordLockedException();
+
         _context.Pages.Add(page);
         await _context.SaveChangesAsync();
         return page;
     }
 
+    /// <summary>
+    /// Copies all pages with a given section id, to a new section id. 
+    /// Used when creating a new form version from an old one. 
+    /// </summary>
+    /// <param name="oldSectionId"></param>
+    /// <param name="newSectionId"></param>
+    /// <returns></returns>
     public async Task<List<Page>> CopyPagesForNewSection(Guid oldSectionId, Guid newSectionId)
     {
         var pagesToMigrate = await GetPagesForSectionAsync(oldSectionId);
@@ -41,21 +79,45 @@ public class PageRepository : IPageRepository
         return pagesToMigrate;
     }
 
-    public async Task<Page?> Update(Page page)
+    /// <summary>
+    /// Updates a page, throws is no page with a given Id is found, 
+    /// or the linked form version isn't in draft. 
+    /// </summary>
+    /// <param name="page"></param>
+    /// <returns></returns>
+    /// <exception cref="RecordNotFoundException"></exception>
+    /// <exception cref="RecordLockedException"></exception>
+    public async Task<Page> Update(Page page)
     {
         var pageToUpdate = await _context.Pages.FirstOrDefaultAsync(v => v.Id == page.Id);
         if (pageToUpdate is null)
-            return null;
+            throw new RecordNotFoundException(page.Id);
+
+        if (!await _context.Sections.AnyAsync(v => v.Id == page.SectionId && v.FormVersion.Status != FormStatus.Draft))
+            throw new RecordLockedException();
+
         pageToUpdate = page;
         await _context.SaveChangesAsync();
         return pageToUpdate;
     }
 
-    public async Task<Page?> Archive(Guid pageId)
+    /// <summary>
+    /// Archives a page, throws is no page with a given Id is found, 
+    /// or the linked form version isn't in draft. 
+    /// </summary>
+    /// <param name="pageId"></param>
+    /// <returns></returns>
+    /// <exception cref="RecordNotFoundException"></exception>
+    /// <exception cref="RecordLockedException"></exception>
+    public async Task<Page> Archive(Guid pageId)
     {
         var pageToUpdate = await _context.Pages.FirstOrDefaultAsync(v => v.Id == pageId);
         if (pageToUpdate is null)
-            return null;
+            throw new RecordNotFoundException(pageId);
+
+        if (!await _context.Sections.AnyAsync(v => v.Id == pageToUpdate.SectionId && v.FormVersion.Status != FormStatus.Draft))
+            throw new RecordLockedException();
+
         _context.Pages.Remove(pageToUpdate);
         await _context.SaveChangesAsync();
         return pageToUpdate;
